Move configs, add configs to run parts of the pipeline and smaller improvements.
