+++ megadepth/configs/hydra/job_logging/custom.yaml
