+++ megadepth/configs/paths.yaml
