+++ megadepth/configs/refinement/pixsfm.yaml
