+++ megadepth/configs/retrieval/netvlad.yaml
